intégration d'un commentaire optionnel pour la réservation. clean des chemin admin/ sur reservations.html.twig
